// src/lib/product/schema.ts
export type Product = {
  active: boolean | null | undefined;
  description: string | null | undefined;
  id: number | null | undefined;
  price: number | null | undefined;
  promotion_price: number | null | undefined;
  quantity_in_stock: number | null | undefined;
  title: string | null | undefined;
};